python/pyproject.toml
[tool.poetry]
name = "fi-instrumentation-otel"
version = "0.1.3"
description = "OpenTelemetry instrumentation"
authors = ["Future AGI <no-reply@futureagi.com>"]
readme = "README.md"

packages = [
    { include = "fi_instrumentation" },
]

exclude = [
    "frameworks",
]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
opentelemetry-api = "^1.29.0"
opentelemetry-sdk = "^1.29.0"
opentelemetry-exporter-otlp-proto-http = "^1.29.0"
opentelemetry-instrumentation = "*"
python-dotenv = "^1.0.0"
requests = "2.32.3"
typing-extensions = "^4.9.0"
pydantic = "^2.9.2"
wrapt = "^1.15.0"
jsonschema = "^4.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
